editor: Malloc check, flush stuck keyboard
The on screen keyboard stays displayed when the editor is closed.

We need to flush the display once more.

This patch adds the display_flush() method to window.c, to execute
wl_display_flush() once, not in a loop like in display_run().

It also adds the NULL checks in constructor and destructor.